Improved approach for handling start and end of daily data.
