+++ app/javascript/react/types/api.ts
@@ -2,7 +2,6 @@ const enum StatusEnum {
   Idle = "IDLE",
   Pending = "PENDING",
   Fulfilled = "FULFILLED",
-  NoData = "NODATA",
   Rejected = "REJECTED",
 }
